change ReverseSequence sequence_lens type to int64

+++ onnx/backend/test/case/node/reversesequence.py
@@ -25,7 +25,7 @@ class ReverseSequence(Base):
                       [1.0, 5.0, 9.0, 13.0],
                       [2.0, 6.0, 10.0, 14.0],
                       [3.0, 7.0, 11.0, 15.0]], dtype=np.float32)
-        sequence_lens = np.array([4, 3, 2, 1], dtype=np.int32)
+        sequence_lens = np.array([4, 3, 2, 1], dtype=np.int64)
 
         y = np.array([[3.0, 6.0, 9.0, 12.0],
                       [2.0, 5.0, 8.0, 13.0],
@@ -48,7 +48,7 @@ class ReverseSequence(Base):
                       [4.0, 5.0, 6.0, 7.0],
                       [8.0, 9.0, 10.0, 11.0],
                       [12.0, 13.0, 14.0, 15.0]], dtype=np.float32)
-        sequence_lens = np.array([1, 2, 3, 4], dtype=np.int32)
+        sequence_lens = np.array([1, 2, 3, 4], dtype=np.int64)
 
         y = np.array([[0.0, 1.0, 2.0, 3.0],
                       [5.0, 4.0, 6.0, 7.0],
